test: split ACP channel transform coverage
